Move types to command files
