+++ commands/app/migrate.ts
@@ -15,8 +15,7 @@ import {
 import { i18n } from '../../lib/lang';
 import { ApiErrorContext, logError } from '../../lib/errorHandlers';
 import { EXIT_CODES } from '../../lib/enums/exitCodes';
-import { MigrateAppOptions } from '../../types/Yargs';
-import { migrateApp2025_2 } from '../../lib/app/migrate';
+import { migrateApp2025_2, MigrateAppArgs } from '../../lib/app/migrate';
 import { uiBetaTag, uiLink } from '../../lib/ui';
 import { migrateApp2023_2 } from '../../lib/app/migrate_legacy';
 
@@ -26,7 +25,7 @@ export const validMigrationTargets = [v2023_2, v2025_2, unstable];
 const command = 'migrate';
 const describe = undefined; // uiBetaTag(i18n(`commands.project.subcommands.migrateApp.header.text.describe`), false);
 
-export async function handler(options: ArgumentsCamelCase<MigrateAppOptions>) {
+export async function handler(options: ArgumentsCamelCase<MigrateAppArgs>) {
   const { derivedAccountId, platformVersion } = options;
   await trackCommandUsage('migrate-app', {}, derivedAccountId);
   const accountConfig = getAccountConfig(derivedAccountId);
@@ -86,7 +85,7 @@ export async function handler(options: ArgumentsCamelCase<MigrateAppOptions>) {
   return process.exit(EXIT_CODES.SUCCESS);
 }
 
-export function builder(yargs: Argv): Argv<MigrateAppOptions> {
+export function builder(yargs: Argv): Argv<MigrateAppArgs> {
   addConfigOptions(yargs);
   addAccountOptions(yargs);
   addUseEnvironmentOptions(yargs);
@@ -125,10 +124,10 @@ export function builder(yargs: Argv): Argv<MigrateAppOptions> {
     ],
   ]);
 
-  return yargs as Argv<MigrateAppOptions>;
+  return yargs as Argv<MigrateAppArgs>;
 }
 
-const migrateCommand: CommandModule<unknown, MigrateAppOptions> = {
+const migrateCommand: CommandModule<unknown, MigrateAppArgs> = {
   command,
   describe,
   handler,
