+++ lib/app/migrate_legacy.ts
@@ -24,12 +24,11 @@ import { trackCommandMetadataUsage } from '../usageTracking';
 import SpinniesManager from '../ui/SpinniesManager';
 import { handleKeypress } from '../process';
 import { poll } from '../polling';
-import { MigrateAppOptions } from '../../types/Yargs';
-import { logInvalidAccountError } from './migrate';
+import { logInvalidAccountError, MigrateAppArgs } from './migrate';
 
 export async function migrateApp2023_2(
   derivedAccountId: number,
-  options: ArgumentsCamelCase<MigrateAppOptions>,
+  options: ArgumentsCamelCase<MigrateAppArgs>,
   accountConfig: CLIAccount
 ): Promise<void> {
   const accountName = uiAccountDescription(derivedAccountId);
